Updated Viewmodel and ViewmodelDataSource
