--- WeatherRxSwift/ViewModel/WeatherDetailViewModel/WeatherDetailDataSource.swift
@@ -1,15 +1,0 @@
-//
-//  WeatherDetailDataSource.swift
-//  WeatherRxSwift
-//
-//  Created by Nischal Hada on 11/8/19.
-//  Copyright © 2019 NischalHada. All rights reserved.
-//
-
-import Foundation
-import RxSwift
-
-protocol WeatherDetailDataSource {
-    var detailList: Observable<[DetailModel]> { get }
-    func getDetailResult()
-}
